refactor: workout drawer components
