+++ convex/workoutSessions.ts
@@ -3,7 +3,7 @@ import { Id } from "./_generated/dataModel";
 import { mutation, MutationCtx } from "./_generated/server";
 import { getCurrentUserOrThrow } from "./users";
 
-const createWorkoutSession = async (ctx: MutationCtx, userId: Id<"users">) => {
+const _createWorkoutSession = async (ctx: MutationCtx, userId: Id<"users">) => {
   const workoutSessionId = await ctx.db.insert("workoutSessions", {
     userId,
     isActive: true,
@@ -16,7 +16,7 @@ export const create = mutation({
   args: {},
   handler: async (ctx) => {
     const user = await getCurrentUserOrThrow(ctx);
-    const workoutSessionId = await createWorkoutSession(ctx, user._id);
+    const workoutSessionId = await _createWorkoutSession(ctx, user._id);
 
     return workoutSessionId;
   },
@@ -37,7 +37,7 @@ export const getOrCreate = mutation({
 
     if (workoutSession) return workoutSession._id;
 
-    const workoutSessionId = await createWorkoutSession(ctx, user._id);
+    const workoutSessionId = await _createWorkoutSession(ctx, user._id);
 
     return workoutSessionId;
   },
@@ -60,6 +60,25 @@ export const setActive = mutation({
   },
 });
 
+export const complete = mutation({
+  args: {
+    workoutSessionId: v.id("workoutSessions"),
+  },
+  handler: async (ctx, args) => {
+    const user = await getCurrentUserOrThrow(ctx);
+    const workoutSession = await ctx.db.get(args.workoutSessionId);
+
+    if (workoutSession?.userId !== user._id) {
+      throw new Error("You are not allowed to update this workout session");
+    }
+
+    await ctx.db.patch(workoutSession._id, {
+      isActive: false,
+      completedAt: Date.now(),
+    });
+  },
+});
+
 export const deleteWorkoutSession = mutation({
   args: {
     id: v.id("workoutSessions"),
@@ -71,6 +90,16 @@ export const deleteWorkoutSession = mutation({
     if (workoutSession?.userId !== user._id) {
       throw new Error("You are not allowed to delete this workout session");
     }
+
+    // Delete all related exercise sets
+    const exerciseSets = await ctx.db
+      .query("exerciseSets")
+      .withIndex("byWorkoutSessionId", (q) => q.eq("workoutSessionId", args.id))
+      .collect();
+
+    await Promise.all(
+      exerciseSets.map((exerciseSet) => ctx.db.delete(exerciseSet._id))
+    );
 
     await ctx.db.delete(workoutSession._id);
   },
