+++ package.json
@@ -12,6 +12,7 @@
     "@clerk/backend": "^2.7.0",
     "@clerk/nextjs": "^6.30.0",
     "@hookform/resolvers": "^5.2.1",
+    "@radix-ui/react-alert-dialog": "^1.1.15",
     "@radix-ui/react-dialog": "^1.1.15",
     "@radix-ui/react-label": "^2.1.7",
     "@radix-ui/react-popover": "^1.1.15",
